+++ shaders/main_files/composite4.glsl
@@ -18,6 +18,14 @@
 #if SSS_PHOSPHOR == 1
 	#include "/lib/super_secret_settings/phosphor.glsl"
 #endif
+#ifdef REFLECTIONS_ENABLED
+	#include "/utils/screen_to_view.glsl"
+	#include "/lib/reflections.glsl"
+#endif
+#if FOG_ENABLED == 1
+	#include "/lib/fog/getFogDistance.glsl"
+	#include "/lib/fog/getFogAmount.glsl"
+#endif
 #if AA_STRATEGY == 1 || AA_STRATEGY == 3 || AA_STRATEGY == 4
 	#include "/lib/fxaa.glsl"
 #endif
@@ -29,6 +37,40 @@
 #endif
 #if SHARPENING_ENABLED == 1
 	#include "/lib/sharpening.glsl"
+#endif
+
+
+
+#ifdef REFLECTIONS_ENABLED
+	void doReflections(inout vec3 color  ARGS_OUT) {
+		
+		// skip sky and fog
+		float depth = texelFetch(DEPTH_BUFFER_ALL, texelcoord, 0).r;
+		float linearDepth = toLinearDepth(depth  ARGS_IN);
+		if (depthIsSky(linearDepth) || depthIsHand(linearDepth)) {return;}
+		
+		// get strengths
+		vec2 reflectionStrengths = texelFetch(REFLECTION_STRENGTH_BUFFER, texelcoord, 0).rg;
+		#if REFLECTIVE_EVERYTHING == 1
+			reflectionStrengths = vec2(1.0, 0.0);
+		#endif
+		if (reflectionStrengths.r + reflectionStrengths.g < 0.01) {return;}
+		
+		// apply fog
+		vec3 viewPos = screenToView(vec3(texcoord, depth)  ARGS_IN);
+		#if FOG_ENABLED == 1
+			#include "/import/gbufferModelViewInverse.glsl"
+			vec3 playerPos = (gbufferModelViewInverse * startMat(viewPos)).xyz;
+			float fogDistance = getFogDistance(playerPos  ARGS_IN);
+			float fogAmount = getFogAmount(fogDistance, playerPos.y  ARGS_IN);
+			reflectionStrengths *= 1.0 - fogAmount;
+		#endif
+		if (reflectionStrengths.r + reflectionStrengths.g < 0.01) {return;}
+		
+		vec3 normal = texelFetch(NORMALS_BUFFER, texelcoord, 0).rgb;
+		addReflection(color, viewPos, normal, MAIN_BUFFER, reflectionStrengths.r, reflectionStrengths.g  ARGS_IN);
+		
+	}
 #endif
 
 void main() {
@@ -46,6 +88,14 @@ void main() {
 	vec3 prev = vec3(0.0);
 	#ifdef DEBUG_OUTPUT_ENABLED
 		vec3 debugOutput = texelFetch(DEBUG_BUFFER, texelcoord, 0).rgb;
+	#endif
+	
+	
+	
+	// ======== REFLECTIONS ========
+	
+	#ifdef REFLECTIONS_ENABLED
+		doReflections(color  ARGS_IN);
 	#endif