+++ shaders/main_files/gbuffers_water.glsl
@@ -108,7 +108,7 @@ void main() {
 	gl_FragData[1] = vec4(normal, 1.0);
 	
 	#if REFLECTIONS_ENABLED == 1
-		/* DRAWBUFFERS:043 */
+		/* DRAWBUFFERS:046 */
 		gl_FragData[2] = vec4(WATER_REFLECTION_STRENGTHS, 0.0, 1.0);
 	#endif