+++ shaders/main_files/gbuffers_weather.glsl
@@ -21,7 +21,7 @@ void main() {
 	gl_FragData[1] = vec4(normal, 1.0);
 	
 	#if REFLECTIONS_ENABLED == 1
-		/* DRAWBUFFERS:043 */
+		/* DRAWBUFFERS:046 */
 		gl_FragData[2] = vec4(0.0, 0.0, 0.0, 1.0);
 	#endif
 	
@@ -43,12 +43,18 @@ void main() {
 void main() {
 	texcoord = (gl_TextureMatrix[0] * gl_MultiTexCoord0).xy;
 	
+	vec4 pos = gl_Vertex;
+	float horizontalAmount = pos.y * WEATHER_HORIZONTAL_AMOUNT * 0.5;
+	pos.x += horizontalAmount;
+	horizontalAmount *= 0.5;
+	pos.z += horizontalAmount;
+	
 	#if ISOMETRIC_RENDERING_ENABLED == 1
 		#include "/import/gbufferModelViewInverse.glsl"
-		vec3 worldPos = endMat(gbufferModelViewInverse * (gl_ModelViewMatrix * gl_Vertex));
+		vec3 worldPos = endMat(gbufferModelViewInverse * (gl_ModelViewMatrix * pos));
 		gl_Position = projectIsometric(worldPos  ARGS_IN);
 	#else
-		gl_Position = ftransform();
+		gl_Position = gl_ProjectionMatrix * gl_ModelViewMatrix * pos;
 	#endif
 	
 	#ifdef TAA_JITTER